--- test/performance/helsinki/data/gbfs.json
@@ -1,43 +1,0 @@
-{
-"last_updated": 1740466606,
-"ttl": 0,
-"version": "2.3",
-"data": {
-"en": {
-"feeds": [
-{
-"name": "system_information",
-"url": "https://mds.bolt.eu/gbfs/2/10/system_information"
-},
-{
-"name": "free_bike_status",
-"url": "file:///home/henrik/hsl/OpenTripPlanner/test/performance/helsinki/data/free_bike_status.json"
-},
-{
-"name": "system_pricing_plans",
-"url": "https://mds.bolt.eu/gbfs/2/10/system_pricing_plans"
-},
-{
-"name": "gbfs_versions",
-"url": "https://mds.bolt.eu/gbfs/2/10/gbfs_versions"
-},
-{
-"name": "vehicle_types",
-"url": "https://mds.bolt.eu/gbfs/2/10/vehicle_types"
-},
-{
-"name": "geofencing_zones",
-"url": "https://mds.bolt.eu/gbfs/2/10/geofencing_zones"
-},
-{
-"name": "station_information",
-"url": "https://mds.bolt.eu/gbfs/2/10/station_information"
-},
-{
-"name": "station_status",
-"url": "https://mds.bolt.eu/gbfs/2/10/station_status"
-}
-]
-}
-}
-}
